Update main.py
I actually realized that the previous method was only working on a single digit separated by delimiter for example "1,2,3" but it failed on "1,22,3". After testing with split method, I realized that this current method is good to go with the current solution. Lets try and see how it will work.