Update main.py
This time the testing process is using python dict to map expected result with the given string of numbers. If there would be a problem then it would tell us. For these test data, the program is working fine but as we can clearly use our instinct that it is not a full fledged solution. So lets work on it further.