Update main.py
According to the question, if the numbers string is empty then it should not proceed and return immediately and in my case, I do not need a return statement in the beginning because if the numbers string is empty it won't be iterated at all. So no need to write extra code.