Update main.py
So, according to the question, I need to check if there is any negative number. So in order to check it, I just need to find the index of the current digit I found in a string then I can check the character just before that digit. If it is a negative sign then the number is negative. The exception will be raised and the whole program will be stopped not moving forward with iteration. The negative numbers are always like this -2 , -5 and etc. So it will not conflict with the delimiter.