Update main.py
So this time, I have added some test cases. Added that I am using python 3.6.9 version(written on the top of the program). Created some separate functions for better readability of the program. I have also added ",0" at the end of the numbers string for better readability of the program.